requirements.txt
langchain
langchain-openai
openai
streamlit
streamlit-chat
unstructured
tiktoken
faiss-cpu
semantic-router
grandalf
flashrank
langchain-google-community
vllm
websocket-client
pillow
python-multipart